pyproject.toml
[project]
name = "tagrefsorter"
description = "CLI tool to normalize LaTeX \\tag numbering in Jupyter Notebook files"
authors = [
    {name = "SuzuSys",email = "sigmatics@outlook.jp"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nbformat (>=5.10.4,<6.0.0)",
    "pylatexenc (>=2.10,<3.0)",
    "markdown-it-py (>=4.0.0,<5.0.0)",
    "mdit-py-plugins (>=0.5.0,<0.6.0)",
]
dynamic = ["version"]

[dependency-groups]
dev = [
    "mypy (>=1.19.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.0,<0.16.0)",
]

[tool.mypy]
mypy_path = ["typings"]


[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
version = "0.0.0"

[tool.poetry.requires-plugins]
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry.scripts]
tagrefsorter = "tagrefsorter.cli:main"
[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Documentation
    "D1",    # undocumented
    "D203",  # one blank line before class
    "D213",  # multi-line summary second line
    
]
unfixable = [
    "F401", # unused import
    "F841", # unused variable
]

logger-objects = ["src.library.logger.LOGGER"]

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",    # assert allowed in tests
    "SLF001",  # Checks for accesses on "private" class members
    "PLR2004", # Magic values allowed in tests
    # "ANN",     # Type annotations optional in tests
    "D",       # Docstrings optional in tests
]